feat(draftjs): Extract functions into utils for other libs to reuse
