+++ src/components/form-elements/draft-js-mention-selector/DraftJSMentionSelectorCore.js
@@ -2,15 +2,17 @@
 import * as React from 'react';
 import { FormattedMessage } from 'react-intl';
 import classNames from 'classnames';
-import { EditorState, Modifier } from 'draft-js';
+import { EditorState } from 'draft-js';
 
 import DatalistItem from '../../datalist-item';
 import DraftJSEditor from '../../draft-js-editor';
 import SelectorDropdown from '../../selector-dropdown';
+import { getActiveMentionForEditorState, addMention } from './utils';
 
 import messages from './messages';
 
 import type { SelectorItems } from '../../../common/types/core';
+import type { Mention } from './utils';
 
 import './MentionSelector.scss';
 
@@ -56,7 +58,7 @@ type Props = {
 };
 
 type State = {
-    activeMention: Object | null,
+    activeMention: Mention | null,
     isFocused: boolean,
     mentionPattern: RegExp,
 };
@@ -110,42 +112,7 @@ class DraftJSMentionSelector extends React.Component<Props, State> {
     getActiveMentionForEditorState(editorState: EditorState) {
         const { mentionPattern } = this.state;
 
-        const contentState = editorState.getCurrentContent();
-        const selectionState = editorState.getSelection();
-
-        const startKey = selectionState.getStartKey();
-        const activeBlock = contentState.getBlockForKey(startKey);
-
-        const cursorPosition = selectionState.getStartOffset();
-
-        let result = null;
-
-        // Break the active block into entity ranges.
-        activeBlock.findEntityRanges(
-            character => character.getEntity() === null,
-            (start, end) => {
-                // Find the active range (is the cursor inside this range?)
-                if (start <= cursorPosition && cursorPosition <= end) {
-                    // Determine if the active range contains a mention.
-                    const activeRangeText = activeBlock.getText().substr(start, cursorPosition - start);
-                    const mentionMatch = activeRangeText.match(mentionPattern);
-
-                    if (mentionMatch) {
-                        result = {
-                            blockID: startKey,
-                            mentionString: mentionMatch[2],
-                            mentionTrigger: mentionMatch[1],
-                            start: start + mentionMatch.index,
-                            end: cursorPosition,
-                        };
-                    }
-                }
-
-                return null;
-            },
-        );
-
-        return result;
+        return getActiveMentionForEditorState(editorState, mentionPattern);
     }
 
     /**
@@ -238,49 +205,8 @@ class DraftJSMentionSelector extends React.Component<Props, State> {
     addMention(mention: Object) {
         const { activeMention } = this.state;
         const { editorState } = this.props;
-        const { start, end } = activeMention || {};
 
-        const { id, name } = mention;
-
-        const contentState = editorState.getCurrentContent();
-        const selectionState = editorState.getSelection();
-
-        const preInsertionSelectionState = selectionState.merge({
-            anchorOffset: start,
-            focusOffset: end,
-        });
-
-        const textToInsert = `@${name}`;
-
-        const contentStateWithEntity = contentState.createEntity('MENTION', 'IMMUTABLE', { id });
-
-        const entityKey = contentStateWithEntity.getLastCreatedEntityKey();
-
-        const contentStateWithLink = Modifier.replaceText(
-            contentState,
-            preInsertionSelectionState,
-            textToInsert,
-            null,
-            entityKey,
-        );
-
-        const spaceOffset = preInsertionSelectionState.getStartOffset() + textToInsert.length;
-        const selectionStateForAddingSpace = preInsertionSelectionState.merge({
-            anchorOffset: spaceOffset,
-            focusOffset: spaceOffset,
-        });
-
-        const contentStateWithLinkAndExtraSpace = Modifier.insertText(
-            contentStateWithLink,
-            selectionStateForAddingSpace,
-            ' ',
-        );
-
-        const editorStateWithLink = EditorState.push(
-            editorState,
-            contentStateWithLinkAndExtraSpace,
-            'change-block-type',
-        );
+        const editorStateWithLink = addMention(editorState, activeMention, mention);
 
         this.setState(
             {
