+++ src/components/form-elements/draft-js-mention-selector/index.js
@@ -1,4 +1,5 @@
 // @flow
+export * from './utils';
 export { default } from './DraftJSMentionSelector';
 export { default as createMentionSelectorState } from './createMentionSelectorState';
 export { default as DraftMentionDecorator } from './DraftMentionDecorator';
